Refactoring models, services to change weather forecast to core project,
